requirements.txt
gunicorn
flask
tensorflow
pillow
numpy
protobuf==3.20.*
tqdm
numpy
pandas 
keras 
opencv-python-headless 
pyngrok
pip
matplotlib
sklearn

